feat: add useCaptain composable
